build fixed; warnings elimnated; test broken

+++ core/BoundedQueue.h
@@ -99,7 +99,7 @@ public:
 	int isvalid() const {return (queuesize==maxsize);}
 	
 	void growTo(int size) {
-	  if (elems.size() < size) {
+	  if ((int)elems.size() < size) {
 	    elems.resize(size);
 	    first=0; maxsize=size; queuesize = 0;last = 0;
 	    for(int i=0;i<size;i++) elems[i]=0;
